policy_demo_python/requirements.txt
aztp-client>=1.0.23
python-dotenv>=1.0.0
termcolor>=1.1.0